+++ Test/SerpentTest.h
@@ -49,7 +49,7 @@ namespace Test
 		virtual std::string Run();
         
     private:
-		void CompareMonteCarlo(std::vector<byte> &Key, std::vector<byte> &Input, std::vector<byte> &Output, unsigned int Count = 100);
+		void CompareMonteCarlo(std::vector<byte> &Key, std::vector<byte> &Input, std::vector<byte> &Output, size_t Count = 100);
 		void CompareOutput();
 		void CompareVector(std::vector<byte> &Key, std::vector<byte> &Input, std::vector<byte> &Output);
 		void OnProgress(std::string Data);
