+++ Test/TwofishTest.h
@@ -47,7 +47,7 @@ namespace Test
 		virtual std::string Run();
 
     private:
-		void CompareMonteCarlo(std::vector<byte> &Key, std::vector<byte> &Input, std::vector<byte> &Output, bool Encrypt = true, unsigned int Count = 10000);
+		void CompareMonteCarlo(std::vector<byte> &Key, std::vector<byte> &Input, std::vector<byte> &Output, bool Encrypt = true, size_t Count = 10000);
 		void CompareVector(std::vector<byte> &Key, std::vector<byte> &Input, std::vector<byte> &Output);
 		void Initialize();
 		void OnProgress(std::string Data);
